Allow world readable tmpfiles
See http://trac.buildbot.net/ticket/3576#comment:3 for more details.
Short version is the `acls` option needs to be specified for the mount,
they aren't, so setfacl will fail.

This is a potential security issue, if untrusted people have access to
the temp files. Anything private would be in the jails where only people
with root have access at the moment so it's acceptable.