Set our own limit for ansible-pull
Ansible has a default of "localhost,<hostnames>,127.0.0.1" for the limit
option. On docs, the hostnames includes "docs,docs.buildbot.net". The
"docs" part is the issue. Not including it causes ansible to run
successfully. Using the --limit option causes the default to be
overriden so do that and construct it to drop the "docs" part.

Fixes ticket:3576, I think.