Default authorized key to empty string and check it's not that
Some point after Ansible 2.1, the check changed and now a variable that
has spaces in its value doesn't work. Explicitly compare against the
empty string, and change the default to the empty string, to change the
comparison to work.

Leaving the default value as nothing (e.g. "authorized_key:") has YAML
use an empty dict as the value. Since the authorized_key is expected to
be a string, use that as a default value.